+++ velox/functions/prestosql/registration/StringFunctionsRegistration.cpp
@@ -91,7 +91,6 @@ void registerStringFunctions(const std::string& prefix) {
       Varchar,
       Varchar>({prefix + "split_to_map"});
   VELOX_REGISTER_VECTOR_FUNCTION(udf_concat, prefix + "concat");
-  VELOX_REGISTER_VECTOR_FUNCTION(udf_concat_ws, prefix + "concat_ws");
   VELOX_REGISTER_VECTOR_FUNCTION(udf_replace, prefix + "replace");
   VELOX_REGISTER_VECTOR_FUNCTION(udf_reverse, prefix + "reverse");
   VELOX_REGISTER_VECTOR_FUNCTION(udf_to_utf8, prefix + "to_utf8");
